requirements.txt
django~=2.1.0
pillow
django-disqus
django-froala-editor